Refactor and reorganize to `InngestFunctionReference`
